+++ packages/legacy/core/__tests__/screens/ProofRequestDetails.test.tsx
@@ -6,10 +6,11 @@ import React from 'react'
 import { ConfigurationContext } from '../../App/contexts/configuration'
 import { NetworkProvider } from '../../App/contexts/network'
 import configurationContext from '../contexts/configuration'
-import { useProofRequestTemplates } from '../../verifier/request-templates'
 import ProofRequestDetails from '../../App/screens/ProofRequestDetails'
 import { ProofRequestType, testIdWithKey } from '../../App'
 import { useTemplates, useTemplate } from '../../App/hooks/proof-request-templates'
+import axios from 'axios'
+import { applyTemplateMarkers, useRemoteProofBundleResolver } from '../../App/utils/proofBundle'
 
 jest.mock('react-native-permissions', () => require('react-native-permissions/mock'))
 jest.mock('@react-native-community/netinfo', () => mockRNCNetInfo)
@@ -35,6 +36,9 @@ jest.mock('../../App/hooks/proof-request-templates', () => ({
   useTemplate: jest.fn(),
 }))
 
+jest.mock('axios', () => ({ create: jest.fn() }))
+
+
 const templates = [
   {
     id: 'Aries:5:StudentFullName:0.0.1:indy',
@@ -48,12 +52,19 @@ const templates = [
           schema: 'XUxBrVSALWHLeycAUhrNr9:3:CL:26293:Student Card',
           requestedAttributes: [
             {
-              name: 'student_first_name',
-              restrictions:[{ cred_def_id: 'XUxBrVSALWHLeycAUhrNr9:3:CL:26293:student_card' }],
+              names: ['student_first_name', 'student_last_name'],
+              restrictions: [{ cred_def_id: 'XUxBrVSALWHLeycAUhrNr9:3:CL:26293:student_card' }],
+              devRestrictions: [{ schema_name: 'student_card' }],
+              non_revoked: { to: "@{now}" },
             },
+          ],
+          requestedPredicates: [
             {
-              name: 'student_last_name',
-              restrictions:[{ cred_def_id: 'XUxBrVSALWHLeycAUhrNr9:3:CL:26293:student_card' }],
+              name: 'expiry_date',
+              predicateType: '>=',
+              predicateValue: "@{currentDate(0)}",
+              restrictions: [{ cred_def_id: 'XUxBrVSALWHLeycAUhrNr9:3:CL:26293:student_card' }],
+              devRestrictions: [{ schema_name: 'student_card' }],
             },
           ],
         },
@@ -61,6 +72,9 @@ const templates = [
     },
   }
 ]
+
+// @ts-ignore
+axios.create.mockImplementation(() => ({ get: () => Promise.resolve({ data: templates }) }))
 // @ts-ignore
 useTemplates.mockImplementation(() => templates)
 // @ts-ignore
@@ -83,6 +97,19 @@ describe('ProofRequestDetails Component', () => {
       </ConfigurationContext.Provider>
     )
   }
+
+  test('Proof bundle resolver works correctly', async () => {
+    const resolver = useRemoteProofBundleResolver("http://localhost:3000")
+    const bundle = await resolver.resolve(true)
+    expect((bundle?.[0].payload.data[0] as any).requestedAttributes[0].restrictions.length).toBe(2)
+  })
+
+  test("Template is parsed correctly", async () => {
+    const template = templates[0]
+    const parsedTemplate = applyTemplateMarkers(template)
+    expect(parsedTemplate.payload.data[0].requestedAttributes[0].non_revoked.to).not.toBe("@{now}")
+    expect(parsedTemplate.payload.data[0].requestedPredicates[0].predicateValue.to).not.toBe("@{currentDate(0)}")
+  })
 
   test('Renders correctly', async () => {
     const tree = renderView({ templateId })
